feat(cpp,rust): support macos arm64 (#327)

+++ python/scripts/build-cuda.ps1
@@ -1,8 +1,9 @@
-$ROOT = git rev-parse --show-toplevel
-& $ROOT/cuda/scripts/build-cuda.ps1
-Set-Location $PSScriptRoot/..
-Remove-Item $ROOT/python/pyxis/pyxis_cuda/ptx -Recurse -Force -ErrorAction SilentlyContinue
-New-Item $ROOT/python/pyxis/pyxis_cuda/ptx -ItemType Directory -ErrorAction SilentlyContinue
-Copy-Item $ROOT/cuda/dist/ptx/* ./pyxis/pyxis_cuda/ptx/
-
-Set-Location $ROOT
+if (-not $IsMacOS) {
+    $ROOT = git rev-parse --show-toplevel
+    & $ROOT/cuda/scripts/build-cuda.ps1
+    Set-Location $PSScriptRoot/..
+    Remove-Item $ROOT/python/pyxis/pyxis_cuda/ptx -Recurse -Force -ErrorAction SilentlyContinue
+    New-Item $ROOT/python/pyxis/pyxis_cuda/ptx -ItemType Directory -ErrorAction SilentlyContinue
+    Copy-Item $ROOT/cuda/dist/ptx/* ./pyxis/pyxis_cuda/ptx/
+    Set-Location $ROOT
+}
